Establish single canonical descriptor writer
