+++ package.json
@@ -9,7 +9,7 @@
     "prepare:generated": "npm run media:sync && node scripts/migrate-release-reputation-snapshot.mjs && npm run rdf:deps && npm run rdf:generate && npm run generate",
     "dev": "npm run prepare:generated && ASTRO_TELEMETRY_DISABLED=1 astro dev",
     "check": "npm run validate:release-contract && npm run prepare:generated && npm run validate:query-matrix && ASTRO_TELEMETRY_DISABLED=1 astro check --minimumFailingSeverity warning",
-    "generate": "node scripts/generate-projections.mjs && node scripts/generate-retrieval-projections.mjs",
+    "generate": "node scripts/generate-projections.mjs && node scripts/generate-retrieval-projections.mjs && node scripts/generate-descriptors.mjs",
     "rdf:generate": "node scripts/generate-rdf.mjs",
     "indexnow:prepare": "node scripts/write-indexnow-key.mjs",
     "validate:cloudflare": "python scripts/configure-cloudflare-edge.py --self-test && python scripts/test-cloudflare-subdomain-migration.py",
