+++ scripts/validate-source.mjs
@@ -4,9 +4,9 @@ import {createHash} from 'node:crypto';
 import {assembleCanonicalContent} from './lib/assemble-content.mjs';
 const root=process.cwd(),data=path.join(root,'src/data');const readJson=async p=>JSON.parse(await readFile(path.join(root,p),'utf8')),fail=m=>{throw new Error(m)},arr=v=>Array.isArray(v)?v:(v==null?[]:[v]),id=v=>typeof v==='string'?v:v?.['@id'];
 const release=await readJson('src/data/release.json'),inv=await readJson('src/data/release-invariants.json'),graph=await readJson('src/data/semantic/knowledge-graph.jsonld'),visible=await readJson('src/data/visible-contract.json'),services=await readJson('src/data/service-registry.json'),answers=await readJson('src/data/answer-registry.json'),hf=await readJson('.release/policy/hf-authority-contract.json');
-for(const f of ['CITATION.cff','codemeta.json','src/content-source/000-frontmatter.md','src/layouts/BaseLayout.astro','src/data/semantic/head-ids.json','src/data/semantic/support-ids.json','src/data/semantic/shapes.ttl','src/data/evidence-registry.json','src/data/evidence-snapshot.json','src/data/volatile-facts.json','src/data/render-calibration.json','scripts/generate-retrieval-projections.mjs','scripts/finalize-dist.mjs','scripts/promote-release.mjs','scripts/zenodo_release.py'])await access(path.join(root,f));
+for(const f of ['CITATION.cff','codemeta.json','src/content-source/000-frontmatter.md','src/layouts/BaseLayout.astro','src/data/semantic/head-ids.json','src/data/semantic/support-ids.json','src/data/semantic/shapes.ttl','src/data/evidence-registry.json','src/data/evidence-snapshot.json','src/data/volatile-facts.json','src/data/render-calibration.json','scripts/generate-retrieval-projections.mjs','scripts/generate-descriptors.mjs','scripts/finalize-dist.mjs','scripts/promote-release.mjs','scripts/zenodo_release.py'])await access(path.join(root,f));
 for(const f of ['scripts/generate-v122-overlays.mjs','scripts/finalize-v122-serving.mjs','scripts/rollover-release-date.mjs','scripts/validate-release-controller.mjs','.release/release-request-v1.2.2.json','.release/release-request-v1.2.2.template.json']){try{await access(path.join(root,f));fail(`Obsolete parallel/one-shot release artifact remains: ${f}`)}catch(e){if(e?.code!=='ENOENT')throw e}}
-const packageJson=await readFile(path.join(root,'package.json'),'utf8'),retrievalGenerator=await readFile(path.join(root,'scripts/generate-retrieval-projections.mjs'),'utf8');if(packageJson.includes('generate-v122-overlays')||packageJson.includes('finalize-v122-serving'))fail('Obsolete version-specific generator/finalizer still wired');if(retrievalGenerator.includes("src/data/projections/llms.txt"))fail('Retrieval generator illegally writes canonical llms projection');
+const packageJson=await readFile(path.join(root,'package.json'),'utf8'),baseGenerator=await readFile(path.join(root,'scripts/generate-projections.mjs'),'utf8'),retrievalGenerator=await readFile(path.join(root,'scripts/generate-retrieval-projections.mjs'),'utf8'),descriptorGenerator=await readFile(path.join(root,'scripts/generate-descriptors.mjs'),'utf8'),finalizer=await readFile(path.join(root,'scripts/finalize-dist.mjs'),'utf8');if(packageJson.includes('generate-v122-overlays')||packageJson.includes('finalize-v122-serving'))fail('Obsolete version-specific generator/finalizer still wired');if(!packageJson.includes('generate-projections.mjs && node scripts/generate-retrieval-projections.mjs && node scripts/generate-descriptors.mjs'))fail('Canonical generator order drift');if(retrievalGenerator.includes("src/data/projections/llms.txt"))fail('Retrieval generator illegally writes canonical llms projection');for(const name of ['datapackage.json','croissant.json','dcat.ttl','void.ttl','linkset.json']){if(baseGenerator.includes(`writeFile(path.join(projections,'${name}')`))fail(`Base generator illegally writes descriptor ${name}`);if(finalizer.includes(`writeFile(path.join(dist,'${name}')`))fail(`Finalizer illegally rewrites descriptor ${name}`);if(!descriptorGenerator.includes(`writeFile(path.join(projections,'${name}')`))fail(`Descriptor generator missing ${name}`)}
 if(release.release!==inv.release||release.dateModified!==inv.date)fail('Release/invariant current identity drift');if(!release.medicalReviewedAt)fail('Explicit medicalReviewedAt missing');if(release.dataset.zenodo.conceptDoi===release.dataset.zenodo.versionDoi)fail('Concept DOI collapsed with current Version DOI');
 const nodes=graph['@graph']||[],byId=new Map(nodes.filter(n=>n?.['@id']).map(n=>[n['@id'],n])),person=byId.get(release.primaryEntity.id),clinic=byId.get(release.clinic.id),dataset=byId.get(release.dataset.id);if(!person||!clinic||!dataset)fail('Person/Clinic/Dataset graph constitution broken');if(id(dataset.creator)!==release.primaryEntity.id||id(dataset.publisher)!==release.primaryEntity.id)fail('Dataset creator/publisher is not the physician');if(!arr(dataset.sameAs).map(id).includes(`https://www.wikidata.org/entity/${release.dataset.wikidata}`))fail('Dataset Wikidata identity missing');if(arr(dataset.sameAs).map(id).some(x=>/github\.com|huggingface\.co|doi\.org|zenodo\.org/.test(x||'')))fail('Source/distribution collapsed into Dataset sameAs');
 const graphIds=new Set(byId.keys());for(const file of ['src/data/semantic/head-ids.json','src/data/semantic/support-ids.json'])for(const ref of await readJson(file))if(!graphIds.has(ref))fail(`${file} references missing graph node ${ref}`);
